Extract tools code
